refactor: establish shared responsive semantics
